requirements.txt
# Deep Learning Frameworks
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10

# PDF Handling
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Model Training & Utilities
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0

# Configuration & Logging
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Pre-trained models and NLP (for text verification)
transformers>=4.30.0
timm>=0.9.0

# Additional utilities
colorama>=0.4.6
